Refactor to fit the requirements.
